fix: split test files and fix GitHub release test
