+++ .github/workflows/release.yaml
@@ -157,6 +157,49 @@ jobs:
         run: |
           ./run_tests.sh -t TestProdBasic
 
+  test_TestThreeBasic:
+    needs:
+      - release
+      - test_TestOneBasic
+    if: needs.release.outputs.release_pr
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+        with:
+          token: ${{secrets.GITHUB_TOKEN}}
+          fetch-depth: 0
+      - id: aws-creds
+        uses: aws-actions/configure-aws-credentials@v4
+        with:
+          role-to-assume: ${{env.AWS_ROLE}}
+          role-session-name: ${{github.run_id}}
+          aws-region: ${{env.AWS_REGION}}
+          role-duration-seconds: 14400 # 4 hours
+          output-credentials: true
+      - name: install-nix
+        run: |
+          curl -L https://nixos.org/nix/install | sh
+          source /home/runner/.nix-profile/etc/profile.d/nix.sh
+          nix --version
+          which nix
+      - name: run_tests
+        shell: '/home/runner/.nix-profile/bin/nix develop --ignore-environment --extra-experimental-features nix-command --extra-experimental-features flakes --keep HOME --keep SSH_AUTH_SOCK --keep IDENTIFIER --keep GITHUB_TOKEN --keep GITHUB_OWNER --keep ZONE --keep AWS_ROLE --keep AWS_REGION --keep AWS_DEFAULT_REGION --keep AWS_ACCESS_KEY_ID --keep AWS_SECRET_ACCESS_KEY --keep AWS_SESSION_TOKEN --keep UPDATECLI_GPGTOKEN --keep UPDATECLI_GITHUB_TOKEN --keep UPDATECLI_GITHUB_ACTOR --keep GPG_SIGNING_KEY --keep NIX_SSL_CERT_FILE --keep NIX_ENV_LOADED --keep TERM --command bash -e {0}'
+        env:
+          AWS_ACCESS_KEY_ID: ${{ steps.aws-creds.outputs.aws-access-key-id }}
+          AWS_SECRET_ACCESS_KEY: ${{ steps.aws-creds.outputs.aws-secret-access-key }}
+          AWS_SESSION_TOKEN: ${{ steps.aws-creds.outputs.aws-session-token }}
+          AWS_MAX_ATTEMPTS: 100
+          AWS_RETRY_MODE: adaptive
+          GITHUB_TOKEN: ${{secrets.GITHUB_TOKEN}}
+          GITHUB_OWNER: rancher
+          IDENTIFIER: ${{github.run_id}}
+          ZONE: ${{secrets.ZONE}}
+          ACME_SERVER_URL: https://acme-v02.api.letsencrypt.org/directory
+          RANCHER_INSECURE: false
+        run: |
+          ./run_tests.sh -t TestThreeBasic
+
+
   test_TestDownstreamBasic:
     needs:
       - release
@@ -200,11 +243,12 @@ jobs:
         run: |
           ./run_tests.sh -t TestDownstreamBasic
 
-  test_TestDownstreamRoles:
+  test_TestDownstreamSplitrole:
     needs:
       - release
       - test_TestOneBasic
       - test_TestProdBasic
+      - test_TestThreeBasic
       - test_TestDownstreamBasic
     if: needs.release.outputs.release_pr
     runs-on: ubuntu-latest
@@ -242,15 +286,16 @@ jobs:
           ACME_SERVER_URL: https://acme-v02.api.letsencrypt.org/directory
           RANCHER_INSECURE: false
         run: |
-          ./run_tests.sh -t TestDownstreamRoles
+          ./run_tests.sh -t TestDownstreamSplitrole
 
   test_Cleanup:
     needs:
       - release
       - test_TestOneBasic
       - test_TestProdBasic
+      - test_TestThreeBasic
       - test_TestDownstreamBasic
-      - test_TestDownstreamRoles
+      - test_TestDownstreamSplitrole
     if: always() && needs.release.outputs.release_pr
     runs-on: ubuntu-latest
     steps:
@@ -293,8 +338,9 @@ jobs:
       - release
       - test_TestOneBasic
       - test_TestProdBasic
+      - test_TestThreeBasic
       - test_TestDownstreamBasic
-      - test_TestDownstreamRoles
+      - test_TestDownstreamSplitrole
       - test_Cleanup
     if: success() && needs.release.outputs.release_pr #Ensure the test jobs succeeded, and that a release PR was created.
     runs-on: ubuntu-latest
