+++ test/tests/three/basic_test.go
@@ -1,4 +1,4 @@
-package prod
+package one
 
 import (
 	"os"
@@ -12,11 +12,11 @@ import (
 	util "github.com/rancher/terraform-rancher2-aws/test/tests"
 )
 
-func TestProd(t *testing.T) {
+func TestThreeBasic(t *testing.T) {
 	t.Parallel()
 	id := util.GetId()
 	region := util.GetRegion()
-	directory := "prod"
+	directory := "three"
 	owner := "terraform-ci@suse.com"
 	util.SetAcmeServer()
 
@@ -44,9 +44,10 @@ func TestProd(t *testing.T) {
 		os.RemoveAll(testDir)
 		t.Fatalf("Error creating test key pair: %s", err)
 	}
-
-	sshAgent := ssh.SshAgentWithKeyPair(t, keyPair.KeyPair)
+  sshAgent := ssh.SshAgentWithKeyPair(t, keyPair.KeyPair)
 	t.Logf("Key %s created and added to agent", keyPair.Name)
+
+  // here we need to create an object storage to hold the remote state
 
 	// use oldest RKE2, remember it releases much more than Rancher
 	_, _, rke2Version, err := util.GetRke2Releases()
@@ -99,6 +100,7 @@ func TestProd(t *testing.T) {
 		SshAgent:                 sshAgent,
 		Upgrade:                  true,
 	})
+
 	_, err = terraform.InitAndApplyE(t, terraformOptions)
 	if err != nil {
 		t.Log("Test failed, tearing down...")
